Reuse shared logic
